Fix Olympiad Manager Rewards
FIX JoeAlisson/L2jOrg#93

+++ Gameserver/src/main/org.l2j.gameserver/org/l2j/gameserver/network/serverpackets/NpcInfoAbnormalVisualEffect.java
@@ -44,7 +44,7 @@ public class NpcInfoAbnormalVisualEffect extends ServerPacket {
         buffer.writeInt(_npc.getTransformationDisplayId());
 
         final Set<AbnormalVisualEffect> abnormalVisualEffects = _npc.getEffectList().getCurrentAbnormalVisualEffects();
-        buffer.writeInt(abnormalVisualEffects.size());
+        buffer.writeShort(abnormalVisualEffects.size());
         for (AbnormalVisualEffect abnormalVisualEffect : abnormalVisualEffects) {
             buffer.writeShort(abnormalVisualEffect.getClientId());
         }
